Install SIGTERM handler in the supervisor
